Add a chalicelib module and move database logic
